feature: S3C-1640 Add functional tests
