+++ eve/workers/unit_and_feature_tests/run_server_tests.bash
@@ -9,7 +9,7 @@ killandsleep () {
 }
 
 # run backbeat server
-npm start & bash tests/utils/wait_for_local_port.bash 8900 40 && npm run ft_server_test
+npm start & bash tests/utils/wait_for_local_port.bash 8900 40 && npm run $1
 
 killandsleep 8900
