+++ index.js
@@ -4,14 +4,12 @@ const werelogs = require('werelogs');
 
 const runServer = require('./lib/api/BackbeatServer');
 
-const Config = process.env.CI !== 'true' ? require('./conf/Config') :
-    require('./tests/config.json');
-
+const config = require('./conf/Config');
 const Logger = werelogs.Logger;
 
 werelogs.configure({
-    level: Config.log.logLevel,
-    dump: Config.log.dumpLevel,
+    level: config.log.logLevel,
+    dump: config.log.dumpLevel,
 });
 
-runServer(Config, Logger);
+runServer(config, Logger);
