+++ package.json
@@ -8,8 +8,9 @@
     "queue_processor": "node extensions/replication/queueProcessor/task.js",
     "replication_status_processor": "node extensions/replication/replicationStatusProcessor/task.js",
     "test": "mocha --recursive tests/unit",
-    "ft_test": "mocha --recursive $(find tests/functional -name '*.js' ! -name 'BackbeatServer.js')",
-    "ft_server_test": "mocha tests/functional/api/BackbeatServer.js",
+    "ft_test:replication": "mocha --recursive $(find tests/functional/replication -name '*.js')",
+    "ft_test:api:routes": "mocha tests/functional/api/routes.js",
+    "ft_test:api:retry": "mocha tests/functional/api/retry.js",
     "bh_test": "mocha --recursive tests/behavior",
     "lint": "eslint $(git ls-files '*.js')",
     "lint_md": "mdlint $(git ls-files '*.md')",
