+++ tests/functional/utils/fakeLogger.js
@@ -1,0 +1,8 @@
+const fakeLogger = {
+    trace: () => {},
+    error: () => {},
+    info: () => {},
+    debug: () => {},
+};
+
+module.exports = fakeLogger;
